+++ samples/browser/src/main/java/org/mozilla/samples/browser/BrowserFragment.kt
@@ -17,7 +17,9 @@ import mozilla.components.feature.contextmenu.ContextMenuCandidate
 import mozilla.components.feature.contextmenu.ContextMenuFeature
 import mozilla.components.feature.downloads.DownloadsFeature
 import mozilla.components.feature.session.CoordinateScrollingFeature
+import mozilla.components.feature.prompts.PromptFeature
 import mozilla.components.feature.session.SessionFeature
+import mozilla.components.feature.session.WindowFeature
 import mozilla.components.feature.storage.HistoryTrackingFeature
 import mozilla.components.feature.tabs.toolbar.TabsToolbarFeature
 import mozilla.components.feature.toolbar.ToolbarAutocompleteFeature
@@ -34,6 +36,8 @@ class BrowserFragment : Fragment(), BackHandler {
     private lateinit var historyTrackingFeature: HistoryTrackingFeature
     private lateinit var scrollFeature: CoordinateScrollingFeature
     private lateinit var contextMenuFeature: ContextMenuFeature
+    private lateinit var promptFeature: PromptFeature
+    private lateinit var windowFeature: WindowFeature
 
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
         return inflater.inflate(R.layout.fragment_browser, container, false)
@@ -96,6 +100,10 @@ class BrowserFragment : Fragment(), BackHandler {
                 requireContext(),
                 components.tabsUseCases,
                 view))
+
+        promptFeature = PromptFeature(components.sessionManager, requireFragmentManager())
+
+        windowFeature = WindowFeature(components.engine, components.sessionManager)
     }
 
     private fun showTabs() {
@@ -111,20 +119,24 @@ class BrowserFragment : Fragment(), BackHandler {
         super.onStart()
 
         sessionFeature.start()
+        windowFeature.start()
         toolbarFeature.start()
         downloadsFeature.start()
         scrollFeature.start()
         contextMenuFeature.start()
+        promptFeature.start()
     }
 
     override fun onStop() {
         super.onStop()
 
         sessionFeature.stop()
+        windowFeature.stop()
         toolbarFeature.stop()
         downloadsFeature.stop()
         scrollFeature.stop()
         contextMenuFeature.stop()
+        promptFeature.stop()
     }
 
     override fun onBackPressed(): Boolean {